+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Potted pale rose translation
+
 ## [1.5.3] - 2024-12-25
 ### Added
 - Potted pale rose