+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.0] - 2025-04-11
 ### Added
 - Potted cactus flower
 - Potted tall flowers (Rose Bush, Peony, Lilac, Sunflower & Pitcher Plant)